requirements.txt
pyserial>=3.5
pyinstaller>=5.0
qrcode>=7.4
reportlab>=4.0
pandas>=2.0
cryptography>=41.0